+++ src/components/property-information/property-information.jsx
@@ -4,6 +4,9 @@ import ReviewForm from "../review-form/review-form";
 import {offerPropTypes, reviewsPropTypes} from "../../prop-types";
 import {connect} from "react-redux";
 import {getOneHotelReviews} from "../../selectors/reviews/one-hotel-reviews-selector";
+import {withReviewForm} from "../../hocs/with-review-form";
+
+const ReviewFormWrapped = withReviewForm(ReviewForm);
 
 const PropertyInformation = (props) => {
   const offer = props.offer;
@@ -82,8 +85,8 @@ const PropertyInformation = (props) => {
         <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{offerReviews.length}</span></h2>
 
         <Reviews offerReviews={offerReviews} />
-
-        <ReviewForm />
+        <ReviewFormWrapped />
+        {/* <ReviewForm /> */}
       </section>
     </div>
   );
